+++ changelog.md
@@ -1,6 +1,10 @@
 ## General Changes
 - Added new command to lookup nearby homes: `/zhomes nearhomes [<radius>]`
 - Added delay to update messages upon joining (Cuz it was annoying)
+- Improved information from `/zhomes info` command
+- Fix converter for older versions
+- Added a way to export/import data via `/zhomes export` and `/zhomes import <data>`
+- Fix support for Geyser-Standalone
 
 
 **Full Changelog**: https://github.com/yL3oft/zHomes/compare/2.2.0...2.2.1